requirements.txt
pytest==7.4
pytest-html==3.1.1
pytest-reporter-html1==0.9.2
pytest-rerunfailures==15.0
requests==2.28.1
pytest-bdd==5.0.0
psycopg2==2.9.10
python-socketio==5.6.0
pytest-json-report==1.5.0
pytest-xdist==2.5.0
crcmod==1.7
zulu==2.0.0
sqlparse==0.4.4
pyyaml==6.0.1
rstr==3.2.2
gherkin-official==28.0.0